feat(shared mvp): move Rx logic to RxPresenter.kt
