--- feature/tv-shows/src/main/java/com/my/tv_shows/base/BasePresenter.kt
@@ -1,17 +1,0 @@
-package com.my.tv_shows.base
-
-abstract class BasePresenter<T : BaseView> {
-
-    var view: T? = null
-
-    fun onViewAttached(view: T) {
-        this.view = view
-        onViewReady()
-    }
-
-    open fun onViewDetached() {
-        view = null
-    }
-
-    abstract fun onViewReady()
-}
