--- feature/tv-shows/src/main/java/com/my/tv_shows/base/BaseView.kt
@@ -1,3 +1,0 @@
-package com.my.tv_shows.base
-
-interface BaseView
