+++ shared/components/src/main/java/com/my/resources/mvp/BaseView.kt
@@ -1,0 +1,3 @@
+package com.my.resources.mvp
+
+interface BaseView
